subtitle_tool/requirements.txt
# Core dependencies
PyQt6>=6.4.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
librosa>=0.10.0
pydub>=0.25.1
soundfile>=0.12.1

# AI models
audio-separator>=1.2.0
whisper-cpp-python>=0.1.0
whisperx>=3.1.0
transformers>=4.30.0

# Subtitle processing
pysrt>=1.1.2
webvtt-py>=0.4.6

# Translation services
deepl>=1.12.0
googletrans>=4.0.0

# Utilities
requests>=2.28.0
aiohttp>=3.8.0
aiofiles>=23.0.0
tqdm>=4.64.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.6.0
psutil>=5.9.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0